# PNG: 1.7-2.0x speed-up in whole image decompression with libdeflate on Intel/AMD CPUs