… frmts/png/pngdataset.cpp