- When whole image decompression is asked, on non-interleave image of
  type Byte, use a dedicated code-path that can use libdeflate instead of zlib to
  do the deflate decompression. Only enable it on CPUs with SSE2, to be able to apply
  fast filters.
- Present the image as a single-block is it is of type Byte and <=  512x512 pixels.
- MRF driver: make it use PNG driver to benefit from above optimization

Tests dataset creation:
```
gdal_translate autotest/gcore/data/small_world.tif small_world_big.png -r bilinear -outsize 8000 4000
gdal_translate small_world_big.png small_world_big.gpkg -co TILEFORMAT=PNG
gdal_translate small_world_big.png small_world_big_png.mrf -co COMPRESS=PNG
```

Reading in band-separate way
----------------------------

hyperfine --warmup 1  'python -c "from osgeo import gdal; ds = gdal.Open(\"small_world_big.png\"); x=ds.ReadRaster()"'
hyperfine --warmup 1  'python -c "from osgeo import gdal; ds = gdal.Open(\"small_world_big.gpkg\"); x=ds.ReadRaster()"'
hyperfine --warmup 1  'python -c "from osgeo import gdal; ds = gdal.Open(\"small_world_big_png.mrf\"); x=ds.ReadRaster()"'

master:
- PNG:
    - zlib: 554 ms
    - zlib-ng: 440 ms
- GPKG:
    - zlib: 660 ms
    - zlib-ng: 568 ms
- MRF:
    - zlib: 577 ms
    - zlib-ng: 465 ms

with this PR:
- PNG:
    - zlib: 470 ms
    - zlib-ng: 355 ms
    - libdeflate: 330 ms (1.67x speedup w.r.t zlib master)
- GPKG:
    - zlib: 461 ms
    - zlib-ng: 355 ms
    - libdeflate: 330 ms (2.00x speedup w.r.t zlib master)
- MRF:
    - zlib: 450 ms
    - zlib-ng: 337 ms
    - libdeflate: 336 ms (1.71x speedup w.r.t zlib master)

Reading in pixel-interleaved way
--------------------------------

hyperfine --warmup 1  'python -c "from osgeo import gdal; ds = gdal.Open(\"small_world_big.png\"); x=ds.ReadRaster(buf_pixel_space=3, buf_band_space=1)"'
hyperfine --warmup 1  'python -c "from osgeo import gdal; ds = gdal.Open(\"small_world_big.gpkg\"); x=ds.ReadRaster(buf_pixel_space=3, buf_band_space=1)"'
hyperfine --warmup 1  'python -c "from osgeo import gdal; ds = gdal.Open(\"small_world_big_png.mrf\"); x=ds.ReadRaster(buf_pixel_space=3, buf_band_space=1)"'

master:
- PNG:
    - zlib: 550 ms
    - zlib-ng: 436 ms
- GPKG:
    - zlib: 686 ms
    - zlib-ng: 590 ms
- MRF:
    - zlib: 596 ms
    - zlib-ng: 489 ms

with this PR:
- PNG:
    - zlib: 459 ms
    - zlib-ng: 346 ms
    - libdeflate: 321 ms (1.71x speedup w.r.t zlib master)
- GPKG:
    - zlib: 487 ms
    - zlib-ng: 381 ms
    - libdeflate: 359 ms (1.91x speedup w.r.t zlib master)
- MRF:
    - zlib: 472 ms
    - zlib-ng: 361 ms
    - libdeflate: 361 ms (1.65x speedup w.r.t zlib master)