@rouault 2022-12-05
CC @lucianpls