@lucianpls 2022-12-05
@rouault 
I take it that all the other second level drivers like the WMS and ESRIC will also benefit, without needing changes.

Thanks!